requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
six==1.16.0
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.9.0
